requirement.txt
torch
torchvision 
pandas 
pillow 
matplotlib 
tqdm
pynvml
ptflops